requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
customtkinter==5.2.0
darkdetect==0.8.0
docker==6.1.3
idna==3.4
numpy>=1.26.0
packaging==23.1
pandas
pillow>=11.2.1
plotly==5.15.0
psutil==5.9.5
pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
tenacity==8.2.3
tzdata==2023.3
urllib3==2.0.4
websocket-client==1.6.1
cairosvg
# Зависимости для работы с PostgreSQL
psycopg2-binary
SQLAlchemy==2.0.41
alembic==1.12.1
bcrypt==4.0.1  # Для хеширования паролей
